@@ -12,15 +12,15 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "^11.2.14",
-    "@angular/cdk": "^11.2.13",
-    "@angular/common": "^11.2.14",
-    "@angular/compiler": "^11.2.14",
-    "@angular/core": "^11.2.14",
-    "@angular/forms": "^11.2.14",
-    "@angular/platform-browser": "^11.2.14",
-    "@angular/platform-browser-dynamic": "^11.2.14",
-    "@angular/router": "^11.2.14",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@auth0/angular-jwt": "^5.0.2",
     "@types/file-saver": "^2.0.3",
     "@types/google.maps": "^3.47.4",
@@ -36,8 +36,8 @@
     "jwt-decode": "^3.1.2",
     "moment": "^2.29.1",
     "primeflex": "^2.0.0",
-    "primeicons": "^4.1.0",
-    "primeng": "^11.4.4",
+    "primeicons": "^6.0.1",
+    "primeng": "^11.4.5",
     "rxjs": "~6.6.0",
     "save": "^2.4.0",
     "tslib": "^2.2.0",
@@ -50,15 +50,15 @@
     "os": false
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.13",
-    "@angular/cli": "^11.2.13",
-    "@angular/compiler-cli": "^11.2.14",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/jasmine": "^3.7.4",
     "@types/node": "^12.20.52",
     "codelyzer": "^6.0.2",
-    "jasmine-core": "~3.6.0",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.1.0",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
@@ -67,6 +67,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.6.4"
   }
 }
