@@ -137,11 +137,11 @@ export function tokenGetter(): any {
         // allowedDomains: ['18.223.1.99:8080'],
         // disallowedRoutes: ['http://18.223.1.99:8080/onsystem/api/v1/oauth/token']
 
-        // allowedDomains: ['localhost:8080'],
-        // disallowedRoutes: ['http://localhost:8080/oauth/token']
+        allowedDomains: ['localhost:8080'],
+        disallowedRoutes: ['http://localhost:8080/oauth/token']
 
-        allowedDomains: ['api.centralonsystem.com.br'],
-        disallowedRoutes: ['//api.centralonsystem.com.br/onsystem/api/v1/oauth/token']
+        // allowedDomains: ['api.centralonsystem.com.br'],
+        // disallowedRoutes: ['//api.centralonsystem.com.br/onsystem/api/v1/oauth/token']
       }
     }),
     TieredMenuModule,
